consume/requirements.txt
kafka-python
psycopg2-binary
pydantic-settings
